fix: Fix bug in graph_collate: ensure edge_index is long rather than float.

+++ src/luz/datasets.py
@@ -44,7 +44,7 @@ def graph_collate(batch: Iterable[luz.Data]) -> luz.Data:
         elif k == "edge_index":
             kw[k] = torch.cat(
                 [
-                    torch.as_tensor(sample[k] + offset)
+                    torch.as_tensor(sample[k] + offset, dtype=torch.long)
                     for sample, offset in zip(batch, edge_index_offsets)
                 ],
                 dim=1,
